New File
As described in the README this file allows manual adjustment in bulk of a portfolio in CB Pro. It will allow setting all the available assets as determined by what the ticker pulls in the ticker fetch code, see the README. Also this list is in the coretamodule.

This code will take arguments, 1 buy/sell to determine which action to take.
2 The amount to hold in that asset, it will sell down or buy up to that point.
3. The amount above which no action will be taken on that asset.

This code is a hack of the index-balancer.py code from the cbpro-index-fund-algo
repo. It requires the key,b64secret and passphrase to be in the defines.py file.
The hack basically allows the command lines parameters to be taken in and used to manual set the action and amounts.

Code is good for manually balancing a portfolio or for example to liquidate the assets.
Uses GTT Limit orders Good Till Time of 1 hour. If the portfolio still has an asset that is not set right, just retry after one hour.